requirements.txt
plotly==5.18.0
kaleido==0.2.1
pandas==2.2.3
Pillow==11.0.0
